fix: merge root layout into [locale] layout for next/root-params support
next/root-params (Next 16.3) only exposes a dynamic segment's params when
that segment's own layout IS the app's root layout — i.e. the segment
folder must contain (or be) app/layout.tsx. With locale routing structured
as a separate app/layout.tsx wrapping app/[locale]/layout.tsx, [locale]
was never the root, so root-params couldn't resolve it and threw "Export
locale doesn't exist in target module".

Fix: delete app/layout.tsx and fold its content (fonts, Mantine/Appearance/
Config providers, cookie-driven theme resolution) directly into
app/[locale]/layout.tsx, making [locale] the true root layout. Since this
layout now receives params.locale directly as the segment's own defining
layout, it reads the locale from params rather than importing
next/root-params here — root-params was only ever needed to avoid passing
locale down through non-Server-Component boundaries, and every page/layout
below [locale] already receives it via params with no ambiguity, so no
other files needed to change.

This also collapses the two DictionaryProvider instantiations (root layout
sourced it from a cookie, [locale]/layout.tsx sourced it from the URL) into
one: the URL segment is now the sole, canonical source, wrapped once here.

app/page.tsx is also removed as dead code: proxy.ts unconditionally
redirects "/" to "/{locale}" before any request reaches app/page.tsx, and
its content pointed at a stale, no-longer-existent route shape.
