--- apps/web/app/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function Home() {
-	redirect("/dashboard/platform/overview");
-}
